+++ R/layers-core.R
@@ -181,7 +181,7 @@ layer_permute <- function(object, dims, input_shape = NULL,
                           name = NULL, trainable = NULL, weights = NULL) {
   
   create_layer(keras$layers$Permute, object, list(
-    dims = as_integer_tuple(dims),
+    dims = as_integer_tuple(dims, force_tuple = TRUE),
     input_shape = normalize_shape(input_shape),
     batch_input_shape = normalize_shape(batch_input_shape),
     batch_size = as_nullable_integer(batch_size),
@@ -343,10 +343,16 @@ layer_masking <- function(object, mask_value = 0.0, input_shape = NULL,
 #' @family core layers
 #' 
 #' @export
-layer_flatten <- function(object, 
-                          batch_size = NULL, name = NULL, trainable = NULL, weights = NULL) {
+layer_flatten <- function(object, input_shape = NULL, dtype = NULL, 
+                          name = NULL, trainable = NULL, weights = NULL) {
   
-  create_layer(keras$layers$Flatten, object, list())
+  create_layer(keras$layers$Flatten, object, list(
+    input_shape = normalize_shape(input_shape),
+    dtype = dtype,
+    name = name,
+    trainable = trainable,
+    weights = weights
+  ))
   
 }
 
@@ -354,11 +360,13 @@ layer_flatten <- function(object,
 
 
 
-as_integer_tuple <- function(x) {
+as_integer_tuple <- function(x, force_tuple = FALSE) {
   if (is.null(x))
     x
-  else
+  else if (is.list(x) || force_tuple)
     tuple(as.list(as.integer(x)))
+  else
+    as.integer(x)
 }
 
 as_nullable_integer <- function(x) {