+++ R/layers-recurrent.R
@@ -20,13 +20,6 @@
 #'   else a symbolic loop will be used. Unrolling can speed-up a RNN, although 
 #'   it tends to be more memory-intensive. Unrolling is only suitable for short 
 #'   sequences.
-#' @param implementation one of {0, 1, or 2}. If set to 0, the RNN will use an 
-#'   implementation that uses fewer, larger matrix products, thus running faster
-#'   on CPU but consuming more memory. If set to 1, the RNN will use more matrix
-#'   products, but smaller ones, thus running slower (may actually be faster on 
-#'   GPU) while consuming less memory. If set to 2 (LSTM/GRU only), the RNN will
-#'   combine the input gate, the forget gate and the output gate into a single 
-#'   matrix, enabling more time-efficient parallelization on the GPU.
 #' @param kernel_initializer Initializer for the `kernel` weights matrix, used 
 #'   for the linear transformation of the inputs..
 #' @param recurrent_initializer Initializer for the `recurrent_kernel` weights 
@@ -57,7 +50,7 @@
 #'   
 #' @export
 layer_simple_rnn <- function(object, units, activation = "tanh", use_bias = TRUE, 
-                             return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE, implementation = 0L,
+                             return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE,
                              kernel_initializer = "glorot_uniform", recurrent_initializer = "orthogonal", bias_initializer = "zeros", 
                              kernel_regularizer = NULL, recurrent_regularizer = NULL, bias_regularizer = NULL, activity_regularizer = NULL, 
                              kernel_constraint = NULL, recurrent_constraint = NULL, bias_constraint = NULL, 
@@ -72,7 +65,6 @@ layer_simple_rnn <- function(object, units, activation = "tanh", use_bias = TRUE
     go_backwards = go_backwards,
     stateful = stateful,
     unroll = unroll,
-    implementation = as.integer(implementation),
     kernel_initializer = kernel_initializer,
     recurrent_initializer = recurrent_initializer,
     bias_initializer = bias_initializer,
@@ -122,7 +114,7 @@ layer_simple_rnn <- function(object, units, activation = "tanh", use_bias = TRUE
 #'     
 #' @export
 layer_gru <- function(object, units, activation = "tanh", recurrent_activation = "hard_sigmoid", use_bias = TRUE, 
-                      return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE, implementation = 0L,
+                      return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE,
                       kernel_initializer = "glorot_uniform", recurrent_initializer = "orthogonal", bias_initializer = "zeros", 
                       kernel_regularizer = NULL, recurrent_regularizer = NULL, bias_regularizer = NULL, activity_regularizer = NULL, 
                       kernel_constraint = NULL, recurrent_constraint = NULL, bias_constraint = NULL, 
@@ -138,7 +130,6 @@ layer_gru <- function(object, units, activation = "tanh", recurrent_activation =
     go_backwards = go_backwards,
     stateful = stateful,
     unroll = unroll,
-    implementation = as.integer(implementation),
     kernel_initializer = kernel_initializer,
     recurrent_initializer = recurrent_initializer,
     bias_initializer = bias_initializer,
@@ -244,7 +235,7 @@ layer_cudnn_gru <- function(object, units,
 #'     
 #' @export
 layer_lstm <- function(object, units, activation = "tanh", recurrent_activation = "hard_sigmoid", use_bias = TRUE, 
-                       return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE, implementation = 0L,
+                       return_sequences = FALSE, return_state = FALSE, go_backwards = FALSE, stateful = FALSE, unroll = FALSE,
                        kernel_initializer = "glorot_uniform", recurrent_initializer = "orthogonal", bias_initializer = "zeros", 
                        unit_forget_bias = TRUE, kernel_regularizer = NULL, recurrent_regularizer = NULL, bias_regularizer = NULL, 
                        activity_regularizer = NULL, kernel_constraint = NULL, recurrent_constraint = NULL, bias_constraint = NULL, 
@@ -260,7 +251,6 @@ layer_lstm <- function(object, units, activation = "tanh", recurrent_activation 
     go_backwards = go_backwards,
     stateful = stateful,
     unroll = unroll,
-    implementation = as.integer(implementation),
     kernel_initializer = kernel_initializer,
     recurrent_initializer = recurrent_initializer,
     bias_initializer = bias_initializer,