+++ R/preprocessing.R
@@ -256,7 +256,6 @@ text_tokenizer <- function(num_words = NULL, filters = '!"#$%&()*+,-./:;<=>?@[\\
 #' @param x Vector/list of strings, or a generator of strings (for 
 #'   memory-efficiency); Alternatively a list of "sequence" (a sequence is a 
 #'   list of integer word indices).
-#' @param ... Unused
 #'
 #' @note Required before using [texts_to_sequences()], [texts_to_matrix()], or 
 #'   [sequences_to_matrix()].
@@ -264,13 +263,68 @@ text_tokenizer <- function(num_words = NULL, filters = '!"#$%&()*+,-./:;<=>?@[\\
 #' @family text tokenization
 #'   
 #' @export
-fit_text_tokenizer <- function(object, x, ...) {
+fit_text_tokenizer <- function(object, x) {
   tokenizer <- object
   if (is.list(x))
     tokenizer$fit_on_sequences(x)
   else
     tokenizer$fit_on_texts(x)
+  invisible(tokenizer)
 }
+
+
+#' Save a text tokenizer to an external file
+#' 
+#' Enables persistence of text tokenizers alongside saved models.
+#' 
+#' @details 
+#' You should always use the same text tokenizer for training and  
+#' prediction. In many cases however prediction will occur in another
+#' session with a version of the model loaded via [load_model_hdf5()].
+#' 
+#' In this case you need to save the text tokenizer object after training
+#' and then reload it prior to prediction.
+#' 
+#' @param object Text tokenizer fit with [fit_text_tokenizer()]
+#' @param filename File to save/load
+#' 
+#' @family text tokenization
+#'
+#' @examples \dontrun{
+#' 
+#' # vectorize texts then save for use in prediction
+#' tokenizer <- text_tokenizer(num_words = 10000) %>% 
+#' fit_text_tokenizer(tokenizer, texts)
+#' save_text_tokenizer(tokenizer, "tokenizer")
+#' 
+#' # (train model, etc.)
+#' 
+#' # ...later in another session
+#' tokenizer <- load_text_tokenizer("tokenizer")
+#' 
+#' # (use tokenizer to preprocess data for prediction)
+#' 
+#' }
+#' 
+#' @importFrom reticulate py_save_object
+#' @export 
+save_text_tokenizer <- function(object, filename) {
+  py_save_object(object, filename)
+  invisible(object)
+}
+
+
+#' @importFrom reticulate py_load_object
+#' @rdname save_text_tokenizer
+#' @export
+load_text_tokenizer <- function(filename) {
+  py_load_object(filename)
+}
+
+
+
+
+
 
 #' Transform each text in texts in a sequence of integers.
 #'
@@ -504,12 +558,11 @@ generator_next <- function(generator, completed = NULL) {
 #' @param augment Whether to fit on randomly augmented samples
 #' @param rounds If `augment`, how many augmentation passes to do over the data
 #' @param seed random seed.
-#' @param ... Unused
 #' 
 #' @family image preprocessing
 #' 
 #' @export
-fit_image_data_generator <- function(object, x, augment = FALSE, rounds = 1, seed = NULL, ...) {
+fit_image_data_generator <- function(object, x, augment = FALSE, rounds = 1, seed = NULL) {
   generator <- object
   history <- generator$fit(
     x = keras_array(x),